+++ README.md
@@ -50,10 +50,24 @@ Run inference
  ![horse_res](https://user-images.githubusercontent.com/91237924/179361905-44fcd4ac-7a9e-41f0-bd07-b6cf07245c04.jpg)
 
 
- ## 6. Run with webcam
+## 6. Run with webcam
  You can also run the sample with webcam for real-time detection
   ```shell
   $ python python/webcam.py -m yolov7.onnx -i 0
  ```
  
 Tips: you can switch the device name to **"GPU"** to boost the performance.
+
+
+## 7. Run with webcam for an IoT web app to count people with yolov7-tiny model
+This is a feature to run `webcam.py` to render computer vision results in the browser. 
+The idea for web browser access is for remote deployments purposes and for convenience purposes use a browser to monitor results if the remote computer can be accessed via VPN or port forwarding. 
+The web app also contains a rest endpoint for people and video frame rate which can be logged on an external source requesting data via a GET requests to the web app.
+
+*To run IoT app `cd` into `python` and use an additional `arg`
+*`$ python webcam.py -i 0 -m ./models/yolov7-tiny.onnx --use-flask`
+*Open browser dial into `localhost:5000` or the IP of the computer on port 5000 after setting up appropriate firewall rules
+*People count rest API GET request: `http://localhost:5000/people-count/` to log data from external IoT platform
+*FPS rest API GET request: `http://localhost:5000/fps/` to log data from external IoT platform
+*Future testing to include MQTT if desired
+
