--- python/main_iot.py
@@ -1,459 +1,0 @@
-from openvino.runtime import Core
-import cv2
-import argparse
-import numpy as np
-import random
-
-import utils
-from flask import Flask
-import threading, time
-
-# py -3.9 main_iot.py -m ./model/yolov7.onnx
-
-
-class YOLOV7_OPENVINO(object):
-    def __init__(self, model_path):
-        # set the hyperparameters
-        self.classes = [
-            "person",
-            "bicycle",
-            "car",
-            "motorcycle",
-            "airplane",
-            "bus",
-            "train",
-            "truck",
-            "boat",
-            "traffic light",
-            "fire hydrant",
-            "stop sign",
-            "parking meter",
-            "bench",
-            "bird",
-            "cat",
-            "dog",
-            "horse",
-            "sheep",
-            "cow",
-            "elephant",
-            "bear",
-            "zebra",
-            "giraffe",
-            "backpack",
-            "umbrella",
-            "handbag",
-            "tie",
-            "suitcase",
-            "frisbee",
-            "skis",
-            "snowboard",
-            "sports ball",
-            "kite",
-            "baseball bat",
-            "baseball glove",
-            "skateboard",
-            "surfboard",
-            "tennis racket",
-            "bottle",
-            "wine glass",
-            "cup",
-            "fork",
-            "knife",
-            "spoon",
-            "bowl",
-            "banana",
-            "apple",
-            "sandwich",
-            "orange",
-            "broccoli",
-            "carrot",
-            "hot dog",
-            "pizza",
-            "donut",
-            "cake",
-            "chair",
-            "couch",
-            "potted plant",
-            "bed",
-            "dining table",
-            "toilet",
-            "tv",
-            "laptop",
-            "mouse",
-            "remote",
-            "keyboard",
-            "cell phone",
-            "microwave",
-            "oven",
-            "toaster",
-            "sink",
-            "refrigerator",
-            "book",
-            "clock",
-            "vase",
-            "scissors",
-            "teddy bear",
-            "hair drier",
-            "toothbrush",
-        ]
-        self.img_size = (640, 640)
-        self.conf_thres = args.conf
-        self.iou_thres = 0.6
-        self.class_num = 80
-        self.colors = [[random.randint(0, 255) for _ in range(3)] for _ in self.classes]
-        self.stride = [8, 16, 32]
-        self.anchor_list = [
-            [12, 16, 19, 36, 40, 28],
-            [36, 75, 76, 55, 72, 146],
-            [142, 110, 192, 243, 459, 401],
-        ]
-
-        ie = Core()
-        self.model = ie.read_model(model_path)
-        self.compiled_model = ie.compile_model(model=self.model, device_name="CPU")
-        self.input_layer = self.compiled_model.input(0)
-
-    def letterbox(self, img, new_shape=(640, 640), color=(114, 114, 114)):
-        # Resize and pad image while meeting stride-multiple constraints
-        shape = img.shape[:2]  # current shape [height, width]
-        if isinstance(new_shape, int):
-            new_shape = (new_shape, new_shape)
-
-        # Scale ratio (new / old)
-        r = min(new_shape[0] / shape[0], new_shape[1] / shape[1])
-        new_unpad = int(round(shape[1] * r)), int(round(shape[0] * r))
-        dw, dh = new_shape[1] - new_unpad[0], new_shape[0] - new_unpad[1]  # wh padding
-
-        # divide padding into 2 sides
-        dw /= 2
-        dh /= 2
-
-        # resize
-        if shape[::-1] != new_unpad:
-            img = cv2.resize(img, new_unpad, interpolation=cv2.INTER_LINEAR)
-        top, bottom = int(round(dh - 0.1)), int(round(dh + 0.1))
-        left, right = int(round(dw - 0.1)), int(round(dw + 0.1))
-
-        # add border
-        img = cv2.copyMakeBorder(
-            img, top, bottom, left, right, cv2.BORDER_CONSTANT, value=color
-        )
-
-        return img
-
-    def xywh2xyxy(self, x):
-        # Convert nx4 boxes from [x, y, w, h] to [x1, y1, x2, y2] where xy1=top-left, xy2=bottom-right
-        y = np.copy(x)
-        y[:, 0] = x[:, 0] - x[:, 2] / 2  # top left x
-        y[:, 1] = x[:, 1] - x[:, 3] / 2  # top left y
-        y[:, 2] = x[:, 0] + x[:, 2] / 2  # bottom right x
-        y[:, 3] = x[:, 1] + x[:, 3] / 2  # bottom right y
-
-        return y
-
-    def nms(self, prediction, conf_thres, iou_thres):
-        predictions = np.squeeze(prediction[0])
-
-        # Filter out object confidence scores below threshold
-        obj_conf = predictions[:, 4]
-        predictions = predictions[obj_conf > conf_thres]
-        obj_conf = obj_conf[obj_conf > conf_thres]
-
-        # Multiply class confidence with bounding box confidence
-        predictions[:, 5:] *= obj_conf[:, np.newaxis]
-
-        # Get the scores
-        scores = np.max(predictions[:, 5:], axis=1)
-
-        # Filter out the objects with a low score
-        valid_scores = scores > conf_thres
-        predictions = predictions[valid_scores]
-        scores = scores[valid_scores]
-
-        # Get the class with the highest confidence
-        class_ids = np.argmax(predictions[:, 5:], axis=1)
-
-        # Get bounding boxes for each object
-        boxes = self.xywh2xyxy(predictions[:, :4])
-
-        # Apply non-maxima suppression to suppress weak, overlapping bounding boxes
-        # indices = nms(boxes, scores, self.iou_threshold)
-        
-
-        indices = cv2.dnn.NMSBoxes(
-            boxes.tolist(), scores.tolist(), conf_thres, iou_thres
-        )
-
-        if len(indices) > 0:
-            indices = indices.flatten()
-            
-        return boxes[indices], scores[indices], class_ids[indices]
-
-    def clip_coords(self, boxes, img_shape):
-        # Clip bounding xyxy bounding boxes to image shape (height, width)
-        boxes[:, 0].clip(0, img_shape[1])  # x1
-        boxes[:, 1].clip(0, img_shape[0])  # y1
-        boxes[:, 2].clip(0, img_shape[1])  # x2
-        boxes[:, 3].clip(0, img_shape[0])  # y2
-
-    def scale_coords(self, img1_shape, img0_shape, coords, ratio_pad=None):
-        # Rescale coords (xyxy) from img1_shape to img0_shape
-        # gain  = old / new
-        if ratio_pad is None:
-            gain = min(img1_shape[0] / img0_shape[0], img1_shape[1] / img0_shape[1])
-            padding = (img1_shape[1] - img0_shape[1] * gain) / 2, (
-                img1_shape[0] - img0_shape[0] * gain
-            ) / 2
-        else:
-            gain = ratio_pad[0][0]
-            padding = ratio_pad[1]
-        coords[:, [0, 2]] -= padding[0]  # x padding
-        coords[:, [1, 3]] -= padding[1]  # y padding
-        coords[:, :4] /= gain
-        self.clip_coords(coords, img0_shape)
-
-    def sigmoid(self, x):
-        return 1 / (1 + np.exp(-x))
-
-    def plot_one_box(self, x, img, color=None, label=None, line_thickness=None):
-        # Plots one bounding box on image img
-        tl = line_thickness or round(
-            0.002 * (img.shape[0] + img.shape[1]) / 2) + 1  # line/font thickness
-        color = (255, 255, 255)
-        
-        xmin = int(x[0])
-        ymin = int(x[1])
-        xmax = int(x[2])
-        ymax = int(x[3])
-        
-        cv2.rectangle(img, (xmin, ymin), (xmax, ymax), color, thickness=tl, lineType=cv2.LINE_AA)
-
-        (w, h), _ = cv2.getTextSize(
-            f"{label:.2f}", cv2.FONT_HERSHEY_SIMPLEX, 0.5, 1
-        )
-        
-        # Prints the text.
-        cv2.rectangle(
-            img, (xmin, ymin + h + 5), (xmin + w + 5, ymin), (255, 255, 255), -1
-        )
-
-        cv2.putText(
-            img,
-            f"{label:.2f}",
-            (xmin, ymin + h),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            0.5,
-            (0, 0, 0),
-            1,
-        )
-        
-    # text gets garbled with out this deep copy
-    def insert_people_count(self,image, people):
-        
-        cv2.putText(
-            image, 
-            f"People: {people}", 
-            (1, 25), 
-            cv2.FONT_HERSHEY_SIMPLEX, 
-            0.8, 
-            (0, 0, 0), 
-            1
-        )
-
-        cv2.putText(
-            image,
-            f"Fps: {str(round(routes.current_fps, 2))}",
-            (1, 50),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            0.8,
-            (0, 0, 0),
-            1
-        )
-
-        return image
-
-    def draw(self, img, boxinfo):
-        
-        people = 0
-        for xyxy, conf, cls in boxinfo:
-            conf = round(conf,5)
-            
-            if cls == 0: # if person
-                self.plot_one_box(
-                    xyxy,
-                    img,
-                    label=conf,
-                    color=self.colors[int(cls)],
-                    line_thickness=2,
-                )
-                people += 1
-            
-        image = self.insert_people_count(img, people)
-        
-        # web app for IoT
-        routes.net_people_count = people
-        routes.framecopy = image
-
-    def infer(self):
-        anchor = np.array(self.anchor_list).astype(float).reshape(3, -1, 2)
-        area = self.img_size[0] * self.img_size[1]
-        size = [
-            int(area / self.stride[0] ** 2),
-            int(area / self.stride[1] ** 2),
-            int(area / self.stride[2] ** 2),
-        ]
-        feature = [[int(j / self.stride[i]) for j in self.img_size] for i in range(3)]
-
-        video_capture = cv2.VideoCapture(0)
-        frame_number = 0
-        start_time = time.time()
-        
-        while True:
-            success, src_img = video_capture.read()
-            
-            if not success:
-                print("ERROR! No video feed, breaking out of script")
-                break
-
-            src_size = src_img.shape[:2]
-
-            # Preprocessing
-            img = self.letterbox(src_img, self.img_size)
-            img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)  # BGR to RGB
-            img = img.astype(dtype=np.float32)
-            img /= 255.0
-            input_image = np.expand_dims(img.transpose(2, 0, 1), 0)  # NHWC to NCHW
-
-            # Do inference
-            pred = self.compiled_model([input_image])
-
-            # Get the each feature map's output data
-            output = []
-            output.append(
-                self.sigmoid(
-                    pred[self.compiled_model.output(0)].reshape(
-                        -1, size[0] * 3, 5 + self.class_num
-                    )
-                )
-            )
-            output.append(
-                self.sigmoid(
-                    pred[self.compiled_model.output(1)].reshape(
-                        -1, size[1] * 3, 5 + self.class_num
-                    )
-                )
-            )
-            output.append(
-                self.sigmoid(
-                    pred[self.compiled_model.output(2)].reshape(
-                        -1, size[2] * 3, 5 + self.class_num
-                    )
-                )
-            )
-
-            # Postprocessing
-            grid = []
-            for _, f in enumerate(feature):
-                grid.append([[i, j] for j in range(f[0]) for i in range(f[1])])
-
-            result = []
-            for i in range(3):
-                src = output[i]
-                xy = src[..., 0:2] * 2.0 - 0.5
-                wh = (src[..., 2:4] * 2) ** 2
-                dst_xy = []
-                dst_wh = []
-                for j in range(3):
-                    dst_xy.append(
-                        (xy[:, j * size[i] : (j + 1) * size[i], :] + grid[i])
-                        * self.stride[i]
-                    )
-                    dst_wh.append(wh[:, j * size[i] : (j + 1) * size[i], :] * anchor[i][j])
-                src[..., 0:2] = np.concatenate((dst_xy[0], dst_xy[1], dst_xy[2]), axis=1)
-                src[..., 2:4] = np.concatenate((dst_wh[0], dst_wh[1], dst_wh[2]), axis=1)
-                result.append(src)
-
-            results = np.concatenate(result, 1)
-            boxes, scores, class_ids = self.nms(results, self.conf_thres, self.iou_thres)
-            img_shape = input_image.shape[2:]
-            self.scale_coords(img_shape, src_size, boxes)
-
-            stop_time = time.time()
-            total_time = stop_time - start_time
-            frame_number += 1
-            routes.current_fps = round(frame_number / total_time, 5)
-            
-            # Draw the results
-            self.draw(src_img, zip(boxes, scores, class_ids))
-            
-       
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(add_help=False)
-    args = parser.add_argument_group("Options")
-    args.add_argument(
-        "-h", "--help", action="help", help="Show this help message and exit."
-    )
-    args.add_argument(
-        "-i",
-        "--input",
-        required=False,
-        type=str,
-        default=0,
-        help="Required. Path to an image file.",
-    )
-    args.add_argument(
-        "-m",
-        "--model",
-        required=True,
-        type=str,
-        help="Required. Path to an .xml or .onnx file with a trained model.",
-    )
-
-    args.add_argument(
-        "-c",
-        "--conf",
-        required=False,
-        type=float,
-        default=0.5,
-        help="set confidence in predictions",
-    )
-    
-    args.add_argument('--use-flask', default=True, action='store_true')
-    args.add_argument('--no-flask', dest='use-flask', action='store_false')
-    
-    args = parser.parse_args()
-    
-    routes = utils.WebAppUtils()
-
-    # flask web app configs
-    USE_FLASK = args.use_flask
-    print("WEB APP ENABLED IS: ",USE_FLASK)
-    
-    if USE_FLASK:
-        flask_app = Flask(__name__)
-        app = utils.FlaskAppWrapper(flask_app)
-        
-        app.add_endpoint('/favicon.ico/', 'favicon', routes.favicon)
-        app.add_endpoint('/people-count/', 'get_people', routes.get_people)
-        app.add_endpoint('/fps/', 'get_fps', routes.get_fps)
-        app.add_endpoint('/video-feed/', 'video_feed', routes.video_feed)
-        app.add_endpoint('/', 'index', routes.index)
-
-        threaded_flask_app = threading.Thread(target=lambda: app.run(
-            host='0.0.0.0', port=5000, use_reloader=False))
-        
-        threaded_flask_app.setDaemon(True)
-        threaded_flask_app.start()
-    
-    try:
-        yolov7_detector = YOLOV7_OPENVINO(args.model)
-        yolov7_detector.infer()
-        
-    except KeyboardInterrupt:
-        print("trying to exit gracefully")
-        print("last PFS measured at: ", routes.current_fps)
-        cv2.destroyAllWindows()
-        print("Killing Flask App Thread")
-        exit(0)
